Moved initial type assignment to InstantiatedTypesAnalysisScheduler.
